Add an 8-bit colour specifier, use to clear the current buffer

+++ gfx/src/buffer/vertex.rs
@@ -3,8 +3,8 @@ gfx_defines!{
     /// 2D vertex representation
     ///
     vertex Vertex2D {
-        pos:        [f32; 2] = "a_Pos",
-        tex_coord:  [f32; 2] = "a_TexCoord",
-        color:      [f32; 4] = "a_Color",
+        pos:        [f32; 2]    = "a_Pos",
+        tex_coord:  [f32; 2]    = "a_TexCoord",
+        color:      [u8; 4]     = "a_Color",
     }
 }
